requirements.txt
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
opencv-python==4.8.0.74
torch==2.0.1
torchvision==0.15.2
albumentations==1.3.1
scikit-learn==1.2.2
tqdm==4.65.0
PyYAML==6.0
tensorboard==2.13.0
pycocotools==2.0.6
ultralytics==8.0.145
pillow==9.5.0
labelme==5.3.0 